Docu: add "Where dreams go to die"
Another documentary on The Barkley Marathons. This time we're following
an athlete (Gary Robbins) attempting to finish the Barkley Marathons
twice. Like the other documentary on this list, it's very interesting if
you're into pushing the limits of sports and yourself.

I can say fairly that "The Barkley Marathons" is my discovery of 2019.