Documentaries: add Leaving Neverland (part 1).
I don't really know what to think about all of this... It's disgusting
that's for sure, but it seems weird to me that the documentary is
presented in this way as well. As if everybody is able to hide their
emotions and doesn't even blame anyone for what happened? Strange...